Funcionamento de tela de não autorizado

+++ Models/PermissaoUsuario.cs
@@ -7,5 +7,5 @@ public class PermissaoUsuario
 {
     public long Id { get; set; }
     //usuario
-    public long PermissaoId { get; set; }
+    public int PermissaoId { get; set; }
 }
